Initial commit of rfe_calc
This is a tool intended to help make the calculations required
by U.S. amateur radio operators vis-a-vis FCC Bulletin 65.  It
comes with no warranty and does not remove the requirement to
do the calculations for your station.  But it is a tool to help
where it can.

The idea originated from the power density calculator written
by W4/VP9KF. But I wanted someting that would save my answers
and provide multiple band information at the same time.

Signed-off-by: Vernon Mauery <vernon@mauery.com>