Editing transmitter bands implemented
You can now click on the edit link on the transmitter bands
(now renamed to tbands, with antenna bands renamed to abands)
and it pulls up a modal box to edit the band.  Then clicking
on save saves the data to the localstorage and re-renders the
page.  The re-rendering is not quite complete, the results
still don't update, but that should be trivial.

Signed-off-by: Vernon Mauery <vernon@mauery.com>