Add try/catch blocks to the rendering phase
The rendering phase can be easily broken if the data
is not correct.  And this in turn can break the reset
data store link because js execution stops at the first
error.  This basically ignores the rendering errors
and gives a usable page until I can get around to
fixing any issues.

Signed-off-by: Vernon Mauery <vernon@mauery.com>