Change the way select_options works
This changes select_options to use key/value pairs
If you pass in an array, it uses the index of the array
as keys.  For associative arrays, it behaves exactly as
expected.

Also adds array_keys and array_zip to make it easier to
use the select_options function on various types.

Signed-off-by: Vernon Mauery <vernon@mauery.com>