Add import/export/reset capabilities
This adds the ability to import and export as well as
the ability to reset all data.  The importer changed the
refresh function a bit and added some cooler error logging.
The refresh function is used as a sanity check when
importing data.  If it doesn't render all portions of the
page, something must be wrong and it aborts.

Signed-off-by: Vernon Mauery <vernon@mauery.com>