Fix connection and antennaband form parsing
Not sure why I didn't catch this before, but some of the
things coming off the connection edit form really cannot
be strings, so they have to be passed through parseFloat
or they cause math errors in the calc_exposure function.

Other forms may suffer from this problem, but the variables
are used in such a way that the stringiness doesn't affect
them like it does for these connection variables.

Signed-off-by: Vernon Mauery <vernon@mauery.com>