Add some more help contents and move reset-sample link
This adds some more help contents to the about section and the
import/export section.  The about section tells a little more
about where the idea originated.  It also puts an hr between
the text and the license and changes the font of the license to
make it easier to distinguish the two. The import/export section
tells about the newly moved reset to sample data button that
I moved off the about page.

Signed-off-by: Vernon Mauery <vernon@mauery.com>