Add some helpful error messages to the screen
In cases where a transmitter or antenna are created but no
bands have been added, a message to click on 'Add a band' is
displayed.  In the case where a connection is made and no
bands line up between the antenna and transmitter, the results
tab shows a message telling that they need to fix this.

This also makes the add/edit dialog titles a little more
uniform and easier to read.

Signed-off-by: Vernon Mauery <vernon@mauery.com>