Force antennas and transmitters to have unique names
This puts a check in the save function of Antenna and Transmitter
that fails if the new name is not unique.  This made it so we had
to modify the modal dialog stuff a bit, but the behavior of non-
unique antennas and transmitters is undefined, so we really need
to discourage this behavior.

Signed-off-by: Vernon Mauery <vernon@mauery.com>