Use jquery-ui dialog instead of my custom popup
This converts all of the modal popup instances for the
class editors to use the jquery-ui dialog class.  This
has a more polished look and simplifies the code a bit.
It does make the code a little longer, but it separates
the html from the js and feels less hackish.

Signed-off-by: Vernon Mauery <vernon@mauery.com>