Add schema support so we can change modulation names
To make the modulation names more appropriate, we need to add
a schema so they can get updated in the local storage.  This
was no small task, but it was mostly self contained.  At the same
time, I noticed that all the load_*() functions are always getting
passed the same thing, so I gave it a default.

Signed-off-by: Vernon Mauery <vernon@mauery.com>