Simplify form error resetting
There was an existing error that showed if multiple
fields had errors.  Only the first field is reported
but after that is fixed, the error state on that field
never got reset.  This simplifies the reset so it gets
reset just before validation to make sure we start with
a clean slate.

Signed-off-by: Vernon Mauery <vernon@mauery.com>