Forgot a factor of four in min dx equation
e = (p*gf)/(4*pi*dx^2)

was:
dx = sqrt((p*gf)/(pi*e))
now:
dx = sqrt((p*gf)/(4*pi*e))

I noticed this when it was telling me that the minimum
distance to my antenna was 26.11 feet, but was in compliance
when the actual distance was only 20 feet.  The real
miniumum distance should be 13.06 feet.

Signed-off-by: Vernon Mauery <vernon@mauery.com>