Use combobox for band selection
As suggested by Doug Reed (N0NAS), users should be able to type in any
old frequency. The preset ham bands are still there, but a number with
an option suffix in (kMG)Hz, (kMG)c (cycles), or m (meters) can be used
to do a custom band. Listings of bands are sorted by frequency, ASC.

These custom bands are then saved as part of the new schema to make it
easier to match up bands between antennas and trnasmitters.

Signed-off-by: Vernon Mauery <vernon@mauery.com>